Revert "nfc: Add New API to get the list of active NFCEE"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1994000a-d4b9-477a-b048-0a9d4c71ae4d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I73d82f6d6918e307abf0581845ea376a20d63c0f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com