Revert "Remove code related to stolen trace sessions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/35bbfd7a-65d1-42a0-b2bd-17ec116d3c9f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9e1f88cf244f0883aedec2f8f76a4afd6874ed18
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com