Revert "nfc: Add New API to get the list of active NFCEE"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/20425113-94d9-4148-b12a-7662c8e7a374). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1e9f37f4cf9631bcc6d5d22ec0ee9048285c408a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com