Migrate more of the FrameworksCoreTests to specify options in Android.bp
Previous cls already created the newly named modules that include the
options for these tests (aosp/3197594) and converted many of the
TEST_MAPPING files to use them (aosp/3151676).
This CL picks up places that we missed in the first round,
mostly in the "presubmit" group>

Test: atest FrameworksCoreTests_internal_app FrameworksCoreTests_server_power FrameworksCoreTests_android_content_pm_PostSubmit FrameworksCoreTests_android_content_res_PostSubmit FrameworksCoreTests_accessibility
Verifed the packages that were run.

Note: this failed:
   atest FrameworksCoreTests --include-filter com.android.server.power.stats.BstatsCpuTimesValidationTest
So I fixed up the bad test mapping to point to PowerStatsTests, which
worked, but I'm not converting that to test_module_options in this cl

Bug: b/344976510
Test-Mapping-Slo-Bypass-Bug: b/335015078
Change-Id: Ib42f38653d315e469c84241765076b750bd74d6d