Migrate more of the FrameworksCoreTests to specify options in Android.bp
Previous cls already created the newly named modules that include the
options for these tests (aosp/3197594) and converted many of the
TEST_MAPPING files to use them (aosp/3151676).
This CL picks up places that we skipped in the first round because they
weren't in the "presubmit" group or they required special owner
approval.

Test: atest
Bug: b/344976510
Test-Mapping-Slo-Bypass-Bug: b/335015078
Change-Id: Ib42f38653d315e469c84241765076b750bd74d6d