PackageManager: Extract lib not zipaligned correct in 16kB devices
If 16kB app compat is enabled; fallback to extracting native libraries
which are not zipaligned correctly.

This is expected to be the case for apps that have not updated to
support 16kB page size and use extractNativeLibs=false.

For app compat it it necessary to allow installation of these unaligned
4kB apps.

Bug: 339709616
Test: Install apps unzipaligned app from playstore (chic-fil-a, ...)
Change-Id: I30a1bc50bc9dea40baf37e7af296e2a4d8e857db
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>