Revert "Use keyguard OWNERS for clocks/customization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/62ddbfcf-827b-4b16-835e-3431986571b3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icf7f0391c3bdfb20b0ef808998c218f13db1b9b9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com