Fix PackageManager#getPackageArchiveInfo to cache derived fields
After switching from old parser to PackageParser2, getPackageArchiveInfo doesn't call hideAsFinal(), which results in assignDerivedFields2() not called and app info flags not initalized. As a result, appinfo.flags
is always 0. Fix it by calling hideAsFinal() to cache derived fields.

Alternative fix is to call assignDerivedFields2() in hideAsParsed(),
but that seems not what the API wants.

Change-Id: I33c1d8446739d6982df1fbec29e65b1c6c33556d