Revert "Remove @Parameters for tests that don't use them."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b9fcb4ac-8409-4713-b991-526d1b926a34). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I075d95f698ff149897cccaece37f61fa4c6f4627
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com