Revert "Remove @Parameters for tests that don't use them."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c872c529-4c04-4cf7-8b82-ec134f570304). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6dedca5f99ce8aec3b5944c64876c86d8682296a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com