Revert "uinput: add recording for simulating a touchpad connection"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/021d68a5-9dc8-479f-ae1f-3db3a025c717). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If893aca047ec4ed0c9827201f3391146f4a6c272
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com