Revert "Remove @Parameters for tests that don't use them."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/542f08eb-fbb9-4c49-90ca-7e58dc331af8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I59e6e32a29692e979848bf87d384addf81dc4995
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com