Revert "BatteryService: expose capacity level API"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/21d32b94-20cc-48a4-95cc-1fc2e0985168). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 368185107
Change-Id: Ic45af8551c22802813a23100892fedf29310cd90
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com