Screenshot preview changes for 3 button navigation in landscape
For 3 button navigation the screenshot preview's touch regions in
seascape and landscape covered the navigation bar, so when you took a
screenshot you could no longer interact with the navigation bar until
the preview was dismissed.

The padding of the preview was also incorrect for seascape and landscape
if the device did not have a cutout.  This was fixed so that the preview
no longer covers the navigation bar in seascape.

Flag: EXEMPT bugfix
Bug: 352512326
Test: manual, take screenshot in different orientations and interact
    with navigation buttons, check that padding is correct in landscape
    and seascape
Change-Id: I48f11243a0dbc10d2f6b11bd1e4ae0aa3605447e