Revert "Revert "Update boot image and system server profiles [M8..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6bbae563-407d-4d6b-9c08-4de845702fb8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 369373497
Change-Id: I2b2886fd2668641c83645374a34fbf18e1f2c4c3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com