Add ActivityManagerService.getJavaMethodExecutableOffset
Adds an opration to ActivityManagerService for the caller to fetch
information about the native executable of a given Java method. The
opration's access control is limited to the UprobeStats module.

Given a JVM method in the form of fully qualified class name, method
name, and fully qualified parameter list, the opreation returns
information from the ODEX file associated with that method. If the
method isn't precompiled, the operation returns null. However, ART can
be enhanced to support returning infomration about JIT compiled methods
in the future.

Bug: 333751994
Test: TH
Change-Id: I1f2dc3780d1bd2a682c1fd3ec41e5c8d73e96fc2