Always restart the process and retry launching the activity when a DeadObjectException occurs.
Summary: This change addresses an edge case in the restart logic, where the underlying process could get killed during a relaunch attempt. When this happens, a DeadObjectException may be thrown during any client transactions that happen as part of the relaunch process.

Right now, any RemoteException triggers a second try in the same stack; however, that's guaranteed to fail specifically for DeadObjectException because the same remote objects owned by the known-dead process are still used. This means that the activity manager gives up and leaves the user with no activity.

This change adds a second handler that always retries on DeadObjectException instead of giving up, since a second failure in this case doesn't mean anything.

Test: Successful Build on master branch

Change-Id: I8d3e3fa4e810a8e33f1785e5ff2c7917a136eac4
Signed-off-by: Abhishek Gadewar <abhishekgadewar@meta.com>