cec: check new device at active path [1/1]
PD#SWPL-185470
BUG:374218178

Problem:
When a new device is connected, tv may not
check whether it's on the active path and
update the active source.

Solution:
Abandon the old condition of checking whether
the device has been in the device list, and
do the check when its logical address and
physical adderess are not equal with the
current active source.

Verify:
loki

Change-Id: Iff202a8ac745f43d662aec0f4519bbe67861e2b3
Signed-off-by: an.xi <an.xi@amlogic.com>