ArrayUtils: add secure zeroization support
Add methods for allocating non-movable byte and char arrays, and methods
for reliably zeroizing byte and char arrays.

These methods will be used in code that handles the user's lockscreen
credential and the secrets derived or unlocked from it.  These methods
provide the needed low-level primitives, and when possible they will be
accessed through helper classes such as LockscreenCredential.

Bug: 320392352
Test: atest FrameworksUtilTests:ArrayUtilsTest
Change-Id: I51b195dd3a26e6c3074bf70a864b147eea7b18b9