Revert "Add landscape layouts for keyguard sim pin and puk"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0143d72a-0edb-4c8f-801e-03136595aa25). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 371647585
Change-Id: Iad92bcbedac2f88525a2081a67528cc906631ecf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com