Revert "Add landscape layouts for keyguard sim pin and puk"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/478d612b-96a6-447a-a6d0-5f871c81b004). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7db924d5c5d6896aefab747ec5ab14e29de14926
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com