Revert "Add landscape layout for keyguard pattern view"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/478d612b-96a6-447a-a6d0-5f871c81b004). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I73066c437fb61e7d63e3600c9d33bb10667000d6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com