Revert "Polishing for migration to disable Weaver"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dfd19775-81a3-4501-8550-ac3baeff7d90). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9852c480adbe59f1053a529d41e12aca4a16b532
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com