Revert "[framework] Add saving routing option oem extension APIs."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6f23b67f-31e0-4628-a4cf-3eaa267c3f8f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I86c61cf2d5449994959eb774233e256acdac72ca
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com