Revert "Add flagging libraries for trade-in mode."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/53486e40-96cf-4951-8e93-0cf2fd89370f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 371036018
Change-Id: I631df9be9b25ddaae8bbd429fd1e0509e5d1ab79
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com