Revert "Hide capture more option for non full-screen image"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/82f9400d-2725-4889-9a08-25d545fbe118). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idfb756b4120df1ca720bfa630386b0c411e1bd0a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com