Revert "Biometrics: make the handleFailedAttempt method to priva..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a0e93195-3531-4e05-8607-4893e9fe5cad). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7a67e1645d0d8a926dd81e1b25e93a9dbcf8fb72
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com