Revert "Remove dependencies on the 1-variant fallback"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6e757a4f-487e-4923-93e2-bd314640c0c8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8e3e2262c5be273a4d9ed1d8d761b8d813b10819
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com