Revert "Remove dependencies on the 1-variant fallback"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2219144e-55b5-466f-b117-7911724da271). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iab329e441ec4acb92dfd0a6883529600dbfd0e49
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com