Revert "Remove dependencies on the 1-variant fallback"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/96891a41-7f08-4285-90bd-a185f052a26c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I462d1070c09d1dcf3b2a4daf3a316bc4ec380766
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com