Revert "Add ACTIVITY_SECURITY_OWNERS."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2e33a933-6d07-4133-907d-4452b0f19f9f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9e9ec5543adb1a105021de8e51f9750b85d34e48
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com