Revert "Add "smoke test" mode to run-ravenwood-tests.sh"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2a40b8c4-2995-47dd-9cff-ed708815f8d4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic886a9874e7a2054e19a8f534d45dd1e203eb7e9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com