Revert "Split dirty-image-objects file between art and framework"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5db79483-4338-49d9-8334-cd0dda534bc4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8e5bccd64e3191b87679b466d156f2ac7266da45
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com