Revert "Split dirty-image-objects file between art and framework"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/995fc12d-7581-47c5-bcbf-d42da38bc7bd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7508bb6731cb42a63beb50f29d8bddb2b03c18fa
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com