Revert "Compat-framework on ravenwood"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5bcf0265-662b-4f89-8687-1412b4da7ce6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 378981506
Change-Id: Iba3496fdfb537a52d0e25a140e86ba1af6eaa389
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com