Remove useless ZramWriteback setup
The purpose of refreshZramSettings() is to sync persist.sys.zram_enabled
system property with zram_enabled preference so that zram is enabled
from the next boot. Changes on the system property after boot_completed
has no effect on the current boot because zram is only set up by init
and current Android does not disable it once it is enabled.

Starting ZramWriteback job in refreshZramSettings() can end up trying
zram writeback while zram is not enabled if zram_enabled preference
turned on after boot_complete for a experiment.

Starting ZramWriteback job in handleSystemReady() once is enough.

This is also to make further mmd zram integration easier.

Bug: 375432472
Test: manual

Change-Id: Ia61f5633fabc2f0aa7e8acb41bb93f621619f6bb