Revert "Compat-framework on ravenwood"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9314e5e8-1536-46f1-ac53-616a17c18643). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 378974372
Change-Id: I951e941687062a7594661de757f8a3e16c75f164
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com