package: Add a new BLE channel sounding feature flag
Rationale for adding a new feature flag:
Channel sounding is a new ranging technology supported by Bluetooth
spec. Ranging apps might want to filter based on device's support
for this new hardware feature. This is an optional BT feature, so
not all devices supporting Bluetooth may support this particular
feature, so a feature flag will be useful for apps to query/filter
to understand if ranging is possible via channel sounding or not.

Bug: 371095923
Test: Compiles
Change-Id: I364685fd25ac1069390c07faa59b9b95e3b25e44