Revert "Add ACTIVITY_SECURITY_OWNERS."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/66cc9b59-3ecf-42d9-8c9b-21cfae803073). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib51defb4607162969e724545868af16e68766b9b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com