Revert "Fix AE LockTaskMode issue on fold devices"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/33785498-cd61-4690-a389-5f5bbd52cdb2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7cd15058b32087fdb137ffabbbb5c0c684f77504
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com