Add a secure setting to allow multiple volume streams.
Summary: Multiple volume streams are controlled by a config resource that is cached and set in the system. As a result, when the device boots, this value is set forever. This change adds a secure setting that is also checked at boot time and can be used to enable multiple volume streams after a device reboot.

Test: Successful Build on master branch

Change-Id: I10087eeea13605bc832146099b98e27d1e28d0e9
Signed-off-by: Abdelrahman Daim <adaim@meta.com>