Add check to make sure that an app is considered updated only when the removed users contains the current user
Summary: An app should be considered updated only if the removed users array contains
the current user.
This makes sure that in multi user case, if multiple users have the same app installed
already and another user tries to install the same app then it won't be considered an
update and the restore will properly take place.
Currently in these situations, restore does not take place as the system thinks
that this is an app update.

Change-Id: Icacd5675291e5d12e1993f7374a6ef7bed6e5ecd
Signed-off-by: Abhishek Gadewar <abhishekgadewar@meta.com>