Abort restore process early if there are no packages to restore.
Summary: This is a fix for a specific issue where system apps with a certain flag set were being filtered out of the restore process, but the process would still run unnecessarily. This change adds a check to see if there are any packages to restore before proceeding, and if not, exits the restore process early.

Test: Successful Build on master branch

Change-Id: I0c9d9f1e53c789a6e5ba360b255fb4ee11569e8c
Signed-off-by: Abhishek Gadewar <abhishekgadewar@meta.com>