Fixed virtual button animation anomaly
In the full-screen video playback scenario, the power button locks and unlocks the screen, and the inset value of the virtual button is abnormal. When performing the disappearing animation, the initial position and the end position obtained are abnormal. Normally, the disappearing animation should be performed to the right, but now it is performed downward. The reason is that the bottom value of the inset is abnormal. In this scenario, can the virtual button be directly animated with alpha instead of position animation?

Test: m
Bug: 380987666

Change-Id: Id0d781a321dbb9aa7cdeaa0f6841bd9fde34c550
Signed-off-by: zhangzhihao <zhangzhihao7@xiaomi.corp-partner.google.com>