Revert "Fix AE LockTaskMode issue on fold devices"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3d5e3252-2056-4baf-bbb9-a1a4907cb5a2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If2491a35e7baaa70d3bfd2b3320522c8be54d6d4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com