Revert "[Ravenwood] Update system property handling"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7fcec49f-c5ea-4f1a-b187-0675ec12b89c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I53144fb82bbc53477e1e6bc1e1b97db07ff0574d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com