Revert "BatteryManager: update capacity level description"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d8ba67f8-521c-453e-b6e0-64b2ee75270f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4e6ef47535011f4d98f3fc19754eeeeee9ba3d4f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com