Revert "Introduce a system feature annotation processor"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f2c2592e-6c2e-40ff-9b65-dad2454ffc60). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I458d4f3dcc565569dd5c001a5a7960d82fff35bf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com