Add config_virtualDisplaySupportsProtectedBuffers
Summary :
The parameter controls whether or not the Display.FLAG_SUPPORTS_PROTECTED_BUFFERS flag should be added to secure virtual displays. Current AOSP behavior is to never add this flag, and this configuration makes it always add this flag when the virtual display is marked as secure. This parameter works very similar to config_wifiDisplaySupportsProtectedBuffers which already exists and is used to add the flag to secure wifi displays.

Test : Check the flags of a secure virtual display that is created and ensure that the Display.FLAG_SUPPORTS_PROTECTED_BUFFERS flag exists

Change-Id: I55701388d5e91281f7274dd22e5bbedebb764b8c
Signed-off-by: Abhishek Gadewar <abhishekgadewar@meta.com>