consider permission for access to private displays
Currently live wallpaper is not able to run on any private
display, since Display.hasAccess is queried multiple times
while setting it up, and the live wallpaper does not necessarily
fulfill one of the requirements in it to pass.

As an OEM developer I want to enable a live wallpaper signed with
a platform key to run on displays which are otherwise not accessible
for third party apps (private displays).

As the least intrusive solution I chose to check for
INTERNAL_SYSTEM_WINDOW permissions inside Display.hasAccess. This
permission already has the same effect if used when calling
startActivity for that display.

Test: CtsDisplayTestCases, CtsWindowManagerDeviceTestCases
Change-Id: Ic8ea719bcfd4cd165a518ed10d414fb7a253a5a6
Signed-off-by: Carsten Rohn <carsten.rohn@bmw.de>