MediaMuxer: Add thread safety to muxer calls
The media muxer checks the state and the validity of the native muxer
object before calls to addTrack and release. These checks were not
protected by a mutex, so there could be a case where the native object
could be released after the check in an addTrack. Adding a synchronized
block to make the checks atomic. The phaser will keep track of any
active calls to the native muxer and will ensure that the release
doesn't proceed till the active calls are complete.

Bug: 354135273
Test: atest MediaMuxerTest
Change-Id: I37552c9674078f1b05f5dd3c175291764f26922b