replace PackageMonitor with internal calls to notify OMS to update overlays
Some system apps may restart the process immediately after uninstalling
the update package. Currently, if the OMS happens to receive the
package changes when the process is in the pending start state,
it will not be able to update enabled overlays to the process.

Therefore, to guarantee that the OMS is notified of package changes
before the app process starts, we need a more reliable mechanism.
Consequently, we've opted to replace the PackageMonitor with an internal call.

Bug: 327102015

Test: atest FrameworksServicesTests:OverlayManagerServiceImplTests
Test: Manually update or uninstall a system app that uses a runtime overlay

Change-Id: I1c32aeefc6d0b2938e0d83a781bf36b0064d25b4
Signed-off-by: zhangshuai10 <zhangshuai10@xiaomi.corp-partner.google.com>