Do not rename AtomicFile on failed sync
Only rename the temporary file to the base name in AtomicFile if the
flush and sync succeded, else there may be a incomplete or empty file
with correct name. Normally this sorts out when the file cache is
written by the system, but becomes a problem if there is a power failure
prior to that.

Flush is currently not required, it does nothing, but in case that
implementation is changed.

Add a unit test for the scenario when closing the FileOutputStream prior
to calling finishWrite.

Test: atest android.util.AtomicFileTest
Bug: 380831140

Change-Id: I06dc54d2ace4f8310322dff9739ef78afcab4d7d
Signed-off-by: Ronny Lundstrom <ronny.lundstrom@aptiv.com>