Revert "aconfig: Add min_sdk for NFC module dependencies"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e21201b4-7165-401a-be77-28949f69e90f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 382476040
Change-Id: I2e187134fd804c94c85607b40ed0f97a3993a73c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com