Revert "Add runtime flag to zygote"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/83fd11a2-eb9a-4d32-8160-1eec14206529). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5b0a54fe57fb322b4acd847124be5061c7fd3539
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com