Revert "Add runtime flag to zygote"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f847e535-5ec7-4c41-b6ec-665105e2ebd7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I61ef249e8a95417fbe62077733c1e8fe5a565242
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com