Revert "Allow SystemServer to create VCN service with a service ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4728641f-a6c9-4e34-b351-0fc7d9e2b1a2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7b29767b3435034b577d191b05c088ee1e4a662f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com