Revert "Add marziana@ to wm/ OWNERS"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/db2c174a-633a-46dd-90e2-5befce0efe6c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I261e4d442ff5bf8ec0ab73a4104b37f7dcc26b15
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com