Revert "Avoid set transition ready to false when backgroud activ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4b644b1-d4d7-46f2-bd17-75fc9d09c48b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icf6863a3985779f6dcb90726924e530d7a2b9086
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com