Fix the issue where the application occasionally does not receive display events.
The onUidImportance callback will only be triggered when the UID's importance changes between <= cutpoint and > cutpoint. In this case, the cutpoint is CACHED.
Assume an application's importance changes from FOREGROUND to CACHED. At this point, the display events sent to it will be pending. Then, if its importance changes back to FOREGROUND, because the importance remains <= cutpoint, the onUidImportance callback will not be executed. As a result, the display events will remain pending, and the application will not receive them.

So, we should only set the display event to pending when the importance is less than CACHED.

Change-Id: I78a9a2a542c845530ac58f1ab1c762cbd9046eb3