Revert "Emit trace sections for warmup and benchmark phases"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2cdfaf9b-078b-4410-b9c3-71b6d3ac8d24). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6753f3c67e521a456e85e44a56380cf85b5b2345
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com