Revert "Emit trace sections for warmup and benchmark phases"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/126bc020-d002-43bb-9b2e-3ab7c11b765c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I866a85530a5126036c0dd105c78f1bbf28f9bc08
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com