Revert "Avoid NullPointerException if Location manager is disabled"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fd918c57-72ec-4ac5-ab6a-5d9915488f5d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6c51c114c1c3717b2ec214cdb17a408e2c6963ce
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com