Adding multiuser pendingintent linting to default Android linting
Creates a Lint warning for usages of `PendingIntent.getActivity` where the first argument (`Context`) is a field named `mContext`.

This is a heuristic for identifying calls that may not be multiuser-aware, with the suggested solutions being using `PendingIntent.getActivityAsUser()` or passing in a user-aware context to the method.

Bug: 314301774
Change-Id: Id784435c436e172be99c6e5c390cebe1ddcf6697
Test: atest PendingIntentGetActivityDetectorTest -c
Flags: EXEMPT lint check