Revert "Avoid NullPointerException if Location manager is disabled"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e8e35ee6-7ec8-4317-ac09-798b2233e601). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idbbc2b397b11bc14eccea7ec3bfbacc5ccb9efeb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com