Revert "Emit trace sections for warmup and benchmark phases"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1e3021d4-82f1-4bb7-86e9-85c19aea612c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8a7b774aa12640a5ede9565c71406a778b2c37fd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com