Revert "Preload HttpEngine into the Zygote via an API"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0d329cce-4480-407b-86a4-cd76e2e02223). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5e4aacbea51a3b414e88952fcf175e8cd7ba92a3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com