Revert "jarjar android.aconfig.storage.**"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/05fb2bb8-dd38-4d97-a53a-1b1bda6e6d9b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibfe42aca8925f5390835c004ab3350ab17075276
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com