Revert "Create aconfig flag for enabling SDK system feature cache"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/05fb2bb8-dd38-4d97-a53a-1b1bda6e6d9b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idd072b214c7bc197ad67a9fd44c9546605f12eda
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com