Add try catch blocks to ensure VCN network is always tracked
There have been end users issues indicate that there can be an
untracked VCN network. This VCN network is connected and
validated, but the platform VCN code fails to trakc it with the
mVcns map. This causes the platform VCN code to fail to remove the
"NOT_VCN_MANAGED" capabilities of the underlying networks, and
fail to stop the VCN instance as part of config deprovisioning.

This change ensures that if there is a VCN network, then there must
be a successfully constructed VCN instance and this instance is
tracked by the mVcns map

Bug: 382106220
Test: atest FrameworksVcnTests && atest CtsVcnTestCases
Flag: EXEMPT very low risk bug fix
Change-Id: I19737815dd46324027993656e93a396f7ba23fed