Revert "Emit trace sections for warmup and benchmark phases"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2c0e349a-fe62-4239-a506-3aa6a41722f0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I48dc51a3707539a9a0d5ad5612e3f27a82206853
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com