Revert "Emit trace sections for warmup and benchmark phases"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/61387436-6e49-4be6-8661-fbbd0691bc45). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icab5cdce28f32517e4781c7e5c90f8c6e5007cdb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com