Revert "Emit trace sections for warmup and benchmark phases"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8528fb8d-e42d-47e6-8a02-524bb61505b4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I144f1a7d2f5cce70b329e9ff5b5285dcd82e6014
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com