Fix clearing callback is missing
Basically, pending actions are cleared, the callback is called,
and the registration of the callback is cleared (null is set at checkIfPendingActionsCleared).
However, it is not cleared when a timeout occurs.
The failure to clear the registration of the callback may cause duplicate callback calls.

Add clearing of the callback registration when a timeout occurs.

Change-Id: I8f93d59afdd8a0c5c28f8483c586fe281422819e