Revert "Java and JNI support for native CodecCapabilities."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/72e6add8-065b-495d-a9cd-b7af901cc7e3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 385025289
Bug: 385025289
Change-Id: I10e5877e793d1d292fb9a3d76f6e300baf21b7b3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com