Revert "Java and JNI support for native VideoCapabilities."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4c3f987a-d2fd-4815-a07a-fc4dba973f60). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 385029304
Bug: 385029304
Change-Id: I1231ce35f83e1ec8f53920ce88c3e12089f7d15f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com