Revert "Java and JNI support for native EncoderCapabilities."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dce00993-bfd1-4cf4-a04d-e0c11ac7da80). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 385030748
Bug: 385030748
Change-Id: Ib7229be2943d4f293664c5db3d63c41122623be3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com