Revert "Java and JNI support for native AudioCapabilities."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dce00993-bfd1-4cf4-a04d-e0c11ac7da80). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 385030748
Bug: 385030748
Change-Id: I36a3fb1cca713f60108e6be2188908c73422f05e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com