Revert "[framework] Expose constructor of ApduServiceInfo."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1d3c7bb8-37b2-4680-ac0b-4f6ad26c6191). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I975eb7908bba691bb61a6245e3f725caa309a360
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com