Revert "Java and JNI support for native PerformancePoint."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/68cf0e7b-3064-4ea6-ab2e-e6098303dd06). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 385032678
Change-Id: I6014e833f7b6915fd2298e9d401d6a38e9ea0f94
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com