Revert "Java and JNI support for native CodecCapabilities."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c4f7d2cc-83a7-4115-85ac-76034a45e8ca). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 385032695
Bug: 385032695
Change-Id: I45dea01bbe8f6cef9e76bd4660007079be13e6f5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com