Revert "Java and JNI support for native PerformancePoint."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/54d14672-66c4-486e-8891-d7c2dbd7b018). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 385032472
Change-Id: I9e0b78a7280b2841bb6050ff90d4e174f1d2d84b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com