Revert "Java and JNI support for native CodecCapabilities."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/54d14672-66c4-486e-8891-d7c2dbd7b018). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 385032472
Change-Id: Ia9a97e17f655d4c0c40905384b83190bfae05682
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com