Revert "Java and JNI support for native VideoCapabilities."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/37309e1c-18bc-4397-922d-e41bd2834aa8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 385033163
Bug: 385033163
Change-Id: I564ccf8e116fb471601efded5df4abb44b623f1b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com