Revert "Java and JNI support for native CodecCapabilities."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2c8230de-7e88-4528-974f-9d9790b04643). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 385033986
Change-Id: I2592be8ed94352f153aa9636afdc96ffec46d689
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com