Revert "Java and JNI support for native EncoderCapabilities."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dff54d25-decc-4a8d-a168-123b778ce534). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 385034286
Change-Id: I7763f6098c2fa7d30ea943ec7843dfb2b055e3e1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com