Revert "Java and JNI support for native AudioCapabilities."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dff54d25-decc-4a8d-a168-123b778ce534). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 385034286
Change-Id: I3e9ecb116afc4029c859e83e7a29a116b18e52ff
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com