Revert "Java and JNI support for native AudioCapabilities."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/83649ef1-de93-472c-98f2-52490091f6d1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 385023849
Change-Id: I7ae50ddcaf5643053e2da0ba090ba4758adf1f01
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com