Revert "Java and JNI support for native EncoderCapabilities."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/83649ef1-de93-472c-98f2-52490091f6d1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 385023849
Change-Id: I4455ac7ffbde4777c83c321ed9cc562aa6df6206
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com