ContentResolver javadocs: warn about blocking methods
The methods to get the MIME type, canonicalize, and uncanonicalize a given Url perform synchronous calls to the Content Provider, which could be in another process. As these calls could potentially block the calling thread for up to 23 seconds (current timeout values), they should not be called on the main thread to avoid ANRs and/or deadlocks.

Change-Id: Ib35a1171b35c4080aba107663e525558f4ba3365