Revert "[framework] Expose constructor of ApduServiceInfo."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c714d492-b71b-4bea-a2f5-33ab892ad135). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4707bfad23af349ef6d09c5e6e633bc5b9d0e063
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com