Revert "[framework] Expose constructor of ApduServiceInfo."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/312a5d75-4e2a-4592-8309-ed0724246fb4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id5570321425acd8561981ad66fcd0b256b472ce2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com