Revert "CameraStateMonitor:fix the potential NPE problem."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/750af4c3-bbbd-4cd3-8e6f-7fbc2da1c2cd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8e87c11048316762f8f4c6809ba7c95d03c4799d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com