Revert "nfc(framework): Split non-updatable portion to a separat..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0c335d8a-ebd8-41c1-a361-2e09835169ac). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I401658875c2bedbd1531792c0adb0c504357ef55
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com