Setup ANGLE when it's the native driver
If ANGLE is being used as the OpenGL ES driver, either because the
various flags/settings force the selection or it's the native driver
(ro.hardware.egl=angle), it always needs to be setup. This allows
for configuring ANGLE feature overrides via EGL attributes, using ANGLE
drivers within an installed APK, etc.

Bug: 372694741
Test: Boot the device with ANGLE as the native driver.
Change-Id: I3e5f6ec621569eb76605f55e9236b58a4ebcee08