Revert "Add call to bind for ConscryptPerfTests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4e7e6f9-32c4-4bff-8675-dcfe868bb567). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I95de30aa082ec4c6dc4549055fc95746bac7a6cb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com