Revert "Regenerate LocaleDataTables.cpp due to ICU 76 upgrade"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1d80e55a-2989-4513-8de5-926f9a399735). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I718950d7c59cc2a4ae64f23f4eb2a078ec622a50
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com