Revert "Regenerate LocaleDataTables.cpp due to ICU 76 upgrade"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bef6e165-d649-4b05-9111-3b84d6ce0cb7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8fe255756f4dfd80c6a201a349878f2187dc7613
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com