Revert "Regenerate LocaleDataTables.cpp due to ICU 76 upgrade"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/15cd5dfe-6005-418d-81ad-0bb753e66f04). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie4c5d516ac7fc782560ffc8315f8d898db26d47c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com