Revert "Java and JNI support for native CodecCapabilities."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8f296a5e-9d2b-4396-8679-ecbf56c461b1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 388348924
Bug: 388348924
Change-Id: I89e8accc5b78f8f43f1f57e89c17b5746a4126e7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com