Revert "Java and JNI support for native PerformancePoint."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8f296a5e-9d2b-4396-8679-ecbf56c461b1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 388348924
Bug: 388348924
Change-Id: I3b446516813a69426a44473ba4f8fee045a78649
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com