Revert "Java and JNI support for native CodecCapabilities."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/deb2197a-990a-48d0-9661-49ce44d96b80). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 388359520
Bug: 388359520
Change-Id: I90e6c79193dc98b9ef8c30eb1f81ae05af244214
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com