Revert "Java and JNI support for native CodecCapabilities."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1fdfe3fa-7d25-44b4-a4a1-01b1148e36ea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 388346900
Bug: 388346900
Change-Id: I879dc48b3b4094f01bf963ef4c822348eadd7dcf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com