Revert "Java and JNI support for native EncoderCapabilities."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1fdfe3fa-7d25-44b4-a4a1-01b1148e36ea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 388346900
Bug: 388346900
Change-Id: I5b2324a3c4b34c38ef5d2aaa393f3060a8bb0d58
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com